refactor(tracer)!: move the OTLP exporter under exporters/
`tracer/otlp` sat beside `tracer/exporters` even though OTLP is an exporter,
so anyone looking for exporters would check `exporters/` and not find the one
they actually needed. Now `@tundralibs/tracer/exporters/otlp`.

The split was never about kind — all three implement the same SpanExporter
contract — it is about dependency weight: Console/Memory pull nothing and are
re-exported from the core barrel, while OTLP pulls restler + guardian and must
stay off the core's import graph.

That isolation is preserved and now verified rather than asserted: `deno info`
on the core barrel resolves 245 modules with ZERO restler/guardian reachable,
while the otlp subpath reaches both. The `exporters` barrel deliberately does
not re-export otlp — doing so would silently undo this, since the core barrel
re-exports `exporters`.

BREAKING for anyone already importing `@tundralibs/tracer/otlp`, which is why
it lands now: tracer 0.2.0 shipped without this subpath at all, so nothing
published is affected.
