+++ packages/tracer/README.md
@@ -166,7 +166,7 @@ pulls an HTTP client into a CLI or worker that only creates spans. Wrap it in
 
 ```typescript
 import { BatchSpanProcessor, Tracer } from '@tundralibs/tracer';
-import { OTLPExporter } from '@tundralibs/tracer/otlp';
+import { OTLPExporter } from '@tundralibs/tracer/exporters/otlp';
 
 const tracer = new Tracer({
   serviceName: 'orders',
