+++ packages/tracer/deno.json
@@ -1,13 +1,13 @@
 {
   "name": "@tundralibs/tracer",
   "version": "0.2.0",
-  "description": "Cross-runtime distributed tracing — W3C Trace Context propagation, automatic span nesting via ambient async context, pluggable samplers and exporters",
+  "description": "Cross-runtime distributed tracing \u2014 W3C Trace Context propagation, automatic span nesting via ambient async context, pluggable samplers and exporters",
   "license": "MIT",
   "exports": {
     ".": "./mod.ts",
     "./errors": "./errors/mod.ts",
     "./exporters": "./exporters/mod.ts",
-    "./otlp": "./otlp/mod.ts",
+    "./exporters/otlp": "./exporters/otlp/mod.ts",
     "./types": "./types/mod.ts"
   },
   "imports": {}
