+++ packages/tracer/exporters/mod.ts
@@ -1,7 +1,7 @@
 /**
  * @fileoverview Built-in span exporters.
  *
- * The OTLP exporter lives behind its own subpath (`@tundralibs/tracer/otlp`)
+ * The OTLP exporter lives behind its own subpath (`@tundralibs/tracer/exporters/otlp`)
  * so the core stays dependency-free — importing a tracer does not pull an HTTP
  * client into the graph.
  *
